backend/requirements.txt
fastapi ~= 0.65.2
uvicorn ~= 0.13.4
psycopg2-binary ~= 2.8.6
sqlalchemy ~= 1.4.16
pandas==1.5.3
numpy==1.23.5
python-multipart ~= 0.0.5
python-dotenv ~= 0.15.0
openpyxl ~= 3.0.7